move query to v1
