--- src/query.rs
@@ -1,12 +1,0 @@
-use serde::Deserialize;
-
-#[derive(Deserialize)]
-pub struct ConvertQuery {
-    pub width: Option<u32>,
-    pub height: Option<u32>,
-    #[serde(rename = "asciiString")]
-    pub ascii_string: Option<String>,
-    #[serde(rename = "fontRatio")]
-    pub font_ratio: Option<f64>,
-    pub reverse: Option<bool>,
-}
