+++ src/routes/api/v1/mod.rs
@@ -4,6 +4,18 @@ mod convert_raw;
 use axum::{routing::post, Router};
 use convert::convert;
 use convert_raw::convert_raw;
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+pub struct ConvertQuery {
+    pub width: Option<u32>,
+    pub height: Option<u32>,
+    #[serde(rename = "asciiString")]
+    pub ascii_string: Option<String>,
+    #[serde(rename = "fontRatio")]
+    pub font_ratio: Option<f64>,
+    pub reverse: Option<bool>,
+}
 
 pub fn create_v1_routes() -> Router {
     Router::new()
