+++ src/routes/api/v2/convert_raw.rs
@@ -1,4 +1,4 @@
-use crate::query::ConvertQuery;
+use super::ConvertQuery;
 use axum::{
     extract::{Multipart, Query},
     Json,
